python-ai-service/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
langchain-community
sentence-transformers
requests
wikipedia
faiss-cpu
python-dotenv
pydantic
